pyproject.toml
[project]
name = "very_simple_rag"
version = "0.1.0"
description = "Simple RAG that works stand-alone"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.11.0",
    "huggingface-hub>=0.31.2",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "langchain-huggingface>=0.2.0",
    "llama-cpp-python>=0.3.9",
    "pypdf>=5.4.0",
    "sentence-transformers>=4.1.0",
]
